requirements.txt
configparser
XlsxWriter
Pillow